refactor: split commands.go into domain-specific files
Extract the 1600-line commands.go monolith into 8 focused files:
commands.go (orchestrator), commands_view.go, commands_editor.go,
commands_search.go, commands_palette.go, commands_explorer.go,
commands_git.go, and callbacks.go.

Multi-line inline handlers are extracted into named App methods
(e.g. TriggerAutocomplete, CloseTab, ShowThemePicker, ExplorerNewFile).
Adds editorPathLang() and withEditorLSP() helpers to eliminate repeated
path/lang boilerplate. Stores reg, running, and quitPending on App to
remove closure captures. Simplifies menus.go helpers to use app.reg.
