+++ cmd/ttt/commands.go
@@ -1,1525 +1,25 @@
 package main
 
 import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"strconv"
 	"strings"
+
 	"github.com/eugenioenko/ttt/internal/command"
 	"github.com/eugenioenko/ttt/internal/config"
-	"github.com/eugenioenko/ttt/internal/core/buffer"
-	"github.com/eugenioenko/ttt/internal/core/diff"
-	"github.com/eugenioenko/ttt/internal/git"
-	"github.com/eugenioenko/ttt/internal/github"
 	"github.com/eugenioenko/ttt/internal/ui"
-
-	"github.com/gdamore/tcell/v2"
 )
 
-func registerCommands(reg *command.Registry, app *App, running *bool, quitPending *bool) {
-	registerViewCommands(reg, app)
-	registerEditorCommands(reg, app, running, quitPending)
-	registerSearchCommands(reg, app)
-	registerPaletteCommands(reg, app)
-	registerExplorerCommands(reg, app)
-	registerGitCommands(reg, app)
-	registerWorkspaceCommands(reg, app)
-	registerPRCommands(reg, app)
-	registerWidgetCallbacks(reg, app)
+func registerCommands(app *App) {
+	registerViewCommands(app)
+	registerEditorCommands(app)
+	registerSearchCommands(app)
+	registerPaletteCommands(app)
+	registerExplorerCommands(app)
+	registerGitCommands(app)
+	registerWorkspaceCommands(app)
+	registerPRCommands(app)
+	registerWidgetCallbacks(app)
 	registerEscapeDismissers(app)
 }
-
-func registerViewCommands(reg *command.Registry, app *App) {
-	reg.Register(command.Command{
-		ID: "sidebar.toggle", Title: "Toggle Sidebar",
-		Handler: app.ToggleSidebar,
-	})
-
-	reg.Register(command.Command{
-		ID: "sidebar.explorer", Title: "Show Explorer",
-		Handler: func() {
-			app.explorer.Reload()
-			app.ShowPanel("explorer", app.explorer)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "sidebar.search", Title: "Show Search",
-		Handler: func() {
-			app.ShowPanel("search", app.search)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "sidebar.searchReplace", Title: "Search and Replace in Files",
-		Handler: func() {
-			if app.sidebar.Visible && app.sidebar.ActivePanel == "search" {
-				app.search.ToggleReplaceMode()
-			} else {
-				app.search.SetReplaceMode(true)
-				app.ShowPanel("search", app.search)
-			}
-			app.root.SetFocus(app.search)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "sidebar.changes", Title: "Show Changes",
-		Handler: func() {
-			app.changes.Refresh()
-			app.ShowPanel("changes", app.changes)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "sidebar.wider", Title: "Increase Sidebar Width",
-		Handler: func() {
-			if app.sidebar.Visible {
-				app.SetSidebarWidth(app.splitPanel.DividerPos + 1)
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "sidebar.narrower", Title: "Decrease Sidebar Width",
-		Handler: func() {
-			if app.sidebar.Visible {
-				app.SetSidebarWidth(app.splitPanel.DividerPos - 1)
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "sidebar.focus", Title: "Focus Sidebar",
-		Handler: app.FocusSidebar,
-	})
-
-	reg.Register(command.Command{
-		ID: "panel.toggle", Title: "Toggle Panel",
-		Handler: app.ToggleBottomPanel,
-	})
-
-	reg.Register(command.Command{
-		ID: "panel.focus", Title: "Focus Panel",
-		Handler: func() {
-			if !app.contentSplit.ShowBottom {
-				app.contentSplit.ShowBottom = true
-			}
-			if w := app.bottomPanel.ActiveWidget(); w != nil {
-				app.root.SetFocus(w)
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "terminal.new", Title: "New Terminal",
-		Handler: app.SpawnTerminal,
-	})
-
-	reg.Register(command.Command{
-		ID: "terminal.toggle", Title: "Toggle Terminal",
-		Handler: func() {
-			if !app.contentSplit.ShowBottom {
-				r := app.contentSplit.GetRect()
-				half := r.H / 2
-				if half > r.H-4 {
-					half = r.H - 4
-				}
-				app.contentSplit.BottomH = half
-				app.showTerminalPanel()
-			} else {
-				app.HideBottomPanel()
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "terminal.fullscreen", Title: "Toggle Terminal Fullscreen",
-		Handler: func() {
-			r := app.contentSplit.GetRect()
-			fullH := r.H - 1
-			if app.contentSplit.ShowBottom && app.contentSplit.BottomH >= fullH {
-				app.HideBottomPanel()
-			} else {
-				app.contentSplit.BottomH = fullH
-				app.showTerminalPanel()
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "terminal.closeAll", Title: "Close All Terminals",
-		Handler: app.CloseAllTerminals,
-	})
-
-	reg.Register(command.Command{
-		ID: "view.keyboardTester", Title: "Keyboard Tester",
-		Handler: func() {
-			kt := ui.NewKeyTesterWidget()
-			kt.Borders = app.borders
-			kt.LookupBinding = func(combo string) string {
-				for _, kb := range app.keybindings {
-					if kb.Key == combo {
-						return kb.Command
-					}
-				}
-				return ""
-			}
-			kt.OnDismiss = func() {
-				app.DismissDialog()
-			}
-			app.ShowDialog(kt)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "about", Title: "About ttt",
-		Handler: func() {
-			openURL("https://github.com/eugenioenko/ttt")
-		},
-	})
-}
-
-func registerEditorCommands(reg *command.Registry, app *App, running *bool, quitPending *bool) {
-	reg.Register(command.Command{
-		ID: "editor.focus", Title: "Focus Editor",
-		Handler: func() {
-			app.FocusEditor()
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.autocomplete", Title: "Trigger Autocomplete",
-		Handler: func() {
-			if !app.settings.Autocomplete.Enabled {
-				return
-			}
-			if app.IsAutocompleteActive() {
-				app.DismissAutocomplete()
-				return
-			}
-			path := app.editorGroup.ActiveFilePath()
-			lang := ""
-			if app.editorGroup.Editor != nil && app.editorGroup.Editor.Highlighter != nil {
-				lang = app.editorGroup.Editor.Highlighter.Language()
-			}
-			if lang == "" {
-				app.StatusWarn("No language detected for this file")
-			} else if _, _, ok := app.lspResolve(path, lang); !ok {
-				app.StatusWarn(lang + " language server is not configured. Add it to settings.json under lsp.servers")
-			} else {
-				line, col := app.editorGroup.ActiveCursor()
-				app.RequestCompletions(path, lang, line, col)
-			}
-		},
-	})
-
-	lspAction := func(action func(path, lang string, line, col int)) {
-		path := app.editorGroup.ActiveFilePath()
-		lang := ""
-		if app.editorGroup.Editor != nil && app.editorGroup.Editor.Highlighter != nil {
-			lang = app.editorGroup.Editor.Highlighter.Language()
-		}
-		line, col := app.editorGroup.ActiveCursor()
-		action(path, lang, line, col)
-	}
-
-	reg.Register(command.Command{
-		ID: "editor.hover", Title: "Show Hover",
-		Handler: func() {
-			path := app.editorGroup.ActiveFilePath()
-			lang := ""
-			if app.editorGroup.Editor != nil && app.editorGroup.Editor.Highlighter != nil {
-				lang = app.editorGroup.Editor.Highlighter.Language()
-			}
-			line, col := app.editorGroup.ActiveCursor()
-			ax, ay, _ := app.editorGroup.CursorPosition()
-			app.RequestHover(path, lang, line, col, ax, ay)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.goToDefinition", Title: "Go to Definition",
-		Handler: func() { lspAction(app.RequestDefinition) },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.goToImplementation", Title: "Go to Implementation",
-		Handler: func() { lspAction(app.RequestImplementation) },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.goToTypeDefinition", Title: "Go to Type Definition",
-		Handler: func() { lspAction(app.RequestTypeDefinition) },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.findReferences", Title: "Find All References",
-		Handler: func() { lspAction(app.RequestReferences) },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.rename", Title: "Rename Symbol",
-		Handler: func() {
-			if app.editorGroup.Editor == nil {
-				return
-			}
-			path := app.editorGroup.ActiveFilePath()
-			lang := ""
-			if app.editorGroup.Editor.Highlighter != nil {
-				lang = app.editorGroup.Editor.Highlighter.Language()
-			}
-			line, col := app.editorGroup.ActiveCursor()
-			word := app.wordAtCursor()
-			app.ShowInputDialog("Rename", "New name", word, func(newName string) {
-				if newName != "" && newName != word {
-					app.RequestRename(path, lang, line, col, newName)
-				}
-			})
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.organizeImports", Title: "Source Action: Organize Imports",
-		Handler: func() {
-			path := app.editorGroup.ActiveFilePath()
-			lang := ""
-			if app.editorGroup.Editor != nil && app.editorGroup.Editor.Highlighter != nil {
-				lang = app.editorGroup.Editor.Highlighter.Language()
-			}
-			app.RequestCodeAction(path, lang, "source.organizeImports")
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.fixAll", Title: "Source Action: Fix All",
-		Handler: func() {
-			path := app.editorGroup.ActiveFilePath()
-			lang := ""
-			if app.editorGroup.Editor != nil && app.editorGroup.Editor.Highlighter != nil {
-				lang = app.editorGroup.Editor.Highlighter.Language()
-			}
-			app.RequestCodeAction(path, lang, "source.fixAll")
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.formatDocument", Title: "Source Action: Format Document",
-		Handler: func() {
-			path := app.editorGroup.ActiveFilePath()
-			lang := ""
-			if app.editorGroup.Editor != nil && app.editorGroup.Editor.Highlighter != nil {
-				lang = app.editorGroup.Editor.Highlighter.Language()
-			}
-			app.RequestFormatting(path, lang)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.formatSelection", Title: "Source Action: Format Selection",
-		Handler: func() {
-			if app.editorGroup.Editor == nil {
-				return
-			}
-			path := app.editorGroup.ActiveFilePath()
-			lang := ""
-			if app.editorGroup.Editor.Highlighter != nil {
-				lang = app.editorGroup.Editor.Highlighter.Language()
-			}
-			sel := app.editorGroup.Editor.Selection
-			if sel == nil || !sel.Active {
-				app.RequestFormatting(path, lang)
-				return
-			}
-			start, end := sel.Range(app.editorGroup.Editor.Cursor.Line, app.editorGroup.Editor.Cursor.Col)
-			app.RequestRangeFormatting(path, lang, start.Line, start.Col, end.Line, end.Col)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "tab.next", Title: "Next Tab",
-		Handler: func() { app.editorGroup.NextTab() },
-	})
-
-	reg.Register(command.Command{
-		ID: "tab.prev", Title: "Previous Tab",
-		Handler: func() { app.editorGroup.PrevTab() },
-	})
-
-	reg.Register(command.Command{
-		ID: "tab.close", Title: "Close Tab",
-		Handler: func() {
-			if !app.editorGroup.IsDirty() {
-				app.editorGroup.CloseTab()
-				return
-			}
-			name := app.editorGroup.ActiveFileName()
-			dialog := ui.NewConfirmDialogWidget3(
-				"Save changes to "+name+"?",
-				"Discard", "Cancel", "Save",
-			)
-			dialog.Borders = app.borders
-			dialog.OnButton[0] = func() {
-				app.DismissDialog()
-				app.editorGroup.CloseTab()
-			}
-			dialog.OnButton[1] = func() {
-				app.DismissDialog()
-			}
-			dialog.OnButton[2] = func() {
-				app.DismissDialog()
-				reg.Execute("file.save")
-				app.editorGroup.CloseTab()
-			}
-			dialog.OnDismiss = func() {
-				app.DismissDialog()
-			}
-			app.ShowDialog(dialog)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "tab.closeOthers", Title: "Close Other Tabs",
-		Handler: func() { app.editorGroup.CloseOtherTabs() },
-	})
-
-	reg.Register(command.Command{
-		ID: "tab.closeAll", Title: "Close All Tabs",
-		Handler: func() { app.editorGroup.CloseAllTabs() },
-	})
-
-	reg.Register(command.Command{
-		ID: "file.new", Title: "New File",
-		Handler: func() {
-			app.editorGroup.OpenBuffer("untitled", &buffer.Buffer{Lines: []string{""}})
-			app.root.SetFocus(app.editorGroup)
-		},
-	})
-
-	saveAs := func() {
-		current := app.editorGroup.ActiveFilePath()
-		initial := ""
-		if current != "untitled" {
-			initial = current
-		}
-		app.ShowInputDialog("Save As", "Filename", initial, func(path string) {
-			if path != "" {
-				app.editorGroup.SaveAs(path)
-			}
-		})
-	}
-
-	reg.Register(command.Command{
-		ID: "file.save", Title: "Save File",
-		Handler: func() {
-			path := app.editorGroup.ActiveFilePath()
-			if app.editorGroup.Editor != nil && app.editorGroup.Editor.Highlighter != nil {
-				lang := app.editorGroup.Editor.Highlighter.Language()
-				app.RunCodeActionsOnSave(path, lang)
-				if app.settings.FormatOnSave {
-					app.FormatOnSave(path, lang)
-				}
-			}
-			if !app.editorGroup.Save() {
-				saveAs()
-				return
-			}
-			path = app.editorGroup.ActiveFilePath()
-			if app.editorGroup.Editor != nil && app.editorGroup.Editor.Highlighter != nil {
-				lang := app.editorGroup.Editor.Highlighter.Language()
-				text := strings.Join(app.editorGroup.Editor.Buf.Lines, "\n")
-				app.NotifyLSPSave(path, lang, text)
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "file.saveAs", Title: "Save As...",
-		Handler: saveAs,
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.undo", Title: "Undo",
-		Handler: func() { app.editorGroup.Undo() },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.redo", Title: "Redo",
-		Handler: func() { app.editorGroup.Redo() },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.selectAll", Title: "Select All",
-		Handler: func() { app.editorGroup.SelectAll() },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.copy", Title: "Copy",
-		Handler: func() { app.editorGroup.Copy() },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.cut", Title: "Cut",
-		Handler: func() { app.editorGroup.Cut() },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.paste", Title: "Paste",
-		Handler: func() { app.editorGroup.Paste() },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.moveLineUp", Title: "Move Line Up",
-		Handler: func() { app.editorGroup.MoveLineUp() },
-	})
-	reg.Register(command.Command{
-		ID: "editor.moveLineDown", Title: "Move Line Down",
-		Handler: func() { app.editorGroup.MoveLineDown() },
-	})
-	reg.Register(command.Command{
-		ID: "editor.duplicateLine", Title: "Duplicate Line",
-		Handler: func() { app.editorGroup.DuplicateLine() },
-	})
-	reg.Register(command.Command{
-		ID: "editor.deleteLine", Title: "Delete Line",
-		Handler: func() { app.editorGroup.DeleteLine() },
-	})
-	reg.Register(command.Command{
-		ID: "editor.insertLineBelow", Title: "Insert Line Below",
-		Handler: func() { app.editorGroup.InsertLineBelow() },
-	})
-	reg.Register(command.Command{
-		ID: "editor.insertLineAbove", Title: "Insert Line Above",
-		Handler: func() { app.editorGroup.InsertLineAbove() },
-	})
-	reg.Register(command.Command{
-		ID: "editor.toggleComment", Title: "Toggle Line Comment",
-		Handler: func() { app.editorGroup.ToggleLineComment() },
-	})
-	reg.Register(command.Command{
-		ID: "editor.moveWordLeft", Title: "Move Word Left",
-		Handler: func() { app.editorGroup.MoveWordLeft(false) },
-	})
-	reg.Register(command.Command{
-		ID: "editor.moveWordRight", Title: "Move Word Right",
-		Handler: func() { app.editorGroup.MoveWordRight(false) },
-	})
-	reg.Register(command.Command{
-		ID: "editor.selectWordLeft", Title: "Select Word Left",
-		Handler: func() { app.editorGroup.MoveWordLeft(true) },
-	})
-	reg.Register(command.Command{
-		ID: "editor.selectWordRight", Title: "Select Word Right",
-		Handler: func() { app.editorGroup.MoveWordRight(true) },
-	})
-	reg.Register(command.Command{
-		ID: "editor.deleteWordLeft", Title: "Delete Word Left",
-		Handler: func() { app.editorGroup.DeleteWordLeft() },
-	})
-	reg.Register(command.Command{
-		ID: "editor.deleteWordRight", Title: "Delete Word Right",
-		Handler: func() { app.editorGroup.DeleteWordRight() },
-	})
-
-	reg.Register(command.Command{
-		ID: "multicursor.selectNext", Title: "Add Next Occurrence",
-		Handler: func() { app.editorGroup.SelectNextOccurrence() },
-	})
-	reg.Register(command.Command{
-		ID: "multicursor.selectAll", Title: "Select All Occurrences",
-		Handler: func() { app.editorGroup.SelectAllOccurrences() },
-	})
-	reg.Register(command.Command{
-		ID: "multicursor.undoCursor", Title: "Undo Last Cursor",
-		Handler: func() { app.editorGroup.UndoLastCursor() },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.quit", Title: "Quit",
-		Handler: func() {
-			if !app.editorGroup.AnyDirty() || *quitPending {
-				for _, tt := range app.terminals {
-					tt.term.Close()
-				}
-				*running = false
-				return
-			}
-			*quitPending = true
-			app.StatusWarn("Unsaved changes. Press Ctrl+Q again to quit.")
-		},
-	})
-}
-
-func registerSearchCommands(reg *command.Registry, app *App) {
-	reg.Register(command.Command{
-		ID: "search.find", Title: "Find",
-		Handler: func() {
-			dv := app.editorGroup.ActiveDiffWidget()
-			if dv != nil {
-				app.showDiffFindBar(dv)
-				return
-			}
-			findBar := ui.NewFindBarWidget()
-			findBar.Borders = app.borders
-			findBar.OnSearch = func(query string, opts ui.SearchOptions) []ui.FindMatch {
-				matches, err := ui.FindInLines(app.editorGroup.Editor.Buf.Lines, query, opts)
-				if err != nil {
-					app.StatusWarn("Invalid regex: " + err.Error())
-					return nil
-				}
-				app.editorGroup.SetSearch(query, matches)
-				return matches
-			}
-			findBar.OnNavigate = func(match ui.FindMatch) {
-				app.editorGroup.SetSearchActive(findBar.Current)
-				app.editorGroup.Editor.Cursor.Line = match.Line
-				app.editorGroup.Editor.Cursor.Col = match.Col
-				app.editorGroup.ScrollToCursor()
-			}
-			findBar.OnDismiss = func() {
-				app.DismissDialog()
-				app.editorGroup.ClearSearch()
-			}
-			app.ShowFindBar(findBar)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "search.findNext", Title: "Find Next",
-		Handler: func() { app.editorGroup.FindNext() },
-	})
-
-	reg.Register(command.Command{
-		ID: "search.findPrev", Title: "Find Previous",
-		Handler: func() { app.editorGroup.FindPrev() },
-	})
-
-	reg.Register(command.Command{
-		ID: "search.clearFind", Title: "Clear Find Highlights",
-		Handler: func() { app.editorGroup.ClearSearch() },
-	})
-
-	reg.Register(command.Command{
-		ID: "search.replace", Title: "Find and Replace",
-		Handler: func() {
-			bar := ui.NewReplaceBarWidget()
-			bar.Borders = app.borders
-			bar.OnSearch = func(query string, opts ui.SearchOptions) []ui.FindMatch {
-				matches, err := ui.FindInLines(app.editorGroup.Editor.Buf.Lines, query, opts)
-				if err != nil {
-					app.StatusWarn("Invalid regex: " + err.Error())
-					return nil
-				}
-				app.editorGroup.SetSearch(query, matches)
-				return matches
-			}
-			bar.OnNavigate = func(match ui.FindMatch) {
-				app.editorGroup.SetSearchActive(bar.Current)
-				app.editorGroup.Editor.Cursor.Line = match.Line
-				app.editorGroup.Editor.Cursor.Col = match.Col
-				app.editorGroup.ScrollToCursor()
-			}
-			bar.OnReplace = func(match ui.FindMatch, replacement string) {
-				app.editorGroup.ReplaceMatch(match, replacement)
-			}
-			bar.OnReplaceAll = func(query, replacement string) {
-				app.editorGroup.ReplaceAll(query, replacement)
-			}
-			bar.OnDismiss = func() {
-				app.DismissDialog()
-				app.editorGroup.ClearSearch()
-			}
-			app.ShowDialog(bar)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "search.clear", Title: "Clear Search Results",
-		Handler: func() {
-			app.search.Input.Text = ""
-			app.search.Input.CursorPos = 0
-			app.search.Groups = nil
-			app.search.FlatList = nil
-			app.search.Selected = 0
-			app.search.ScrollTop = 0
-			app.editorGroup.ClearSearch()
-		},
-	})
-}
-
-func registerPaletteCommands(reg *command.Registry, app *App) {
-	openPalette := func(fileMode bool, initialText ...string) {
-		palette := ui.NewCommandPaletteWidget(reg.List())
-		palette.Borders = app.borders
-		palette.SetFiles(app.workspace.Paths())
-		if len(initialText) > 0 {
-			palette.Input.SetText(initialText[0])
-		} else if fileMode {
-			palette.Input.SetText("")
-		}
-		palette.OnExecute = func(id string) {
-			app.DismissDialog()
-			reg.Execute(id)
-		}
-		palette.OnOpenFile = func(absPath string) {
-			app.DismissDialog()
-			app.editorGroup.OpenFile(absPath)
-		}
-		palette.OnGoToLine = func(line int) {
-			app.DismissDialog()
-			app.editorGroup.GoToLine(line)
-		}
-		palette.OnDismiss = func() {
-			app.DismissDialog()
-		}
-		app.ShowDialog(palette)
-	}
-
-	reg.Register(command.Command{
-		ID: "command.palette", Title: "Command Palette",
-		Handler: func() { openPalette(false) },
-	})
-
-	reg.Register(command.Command{
-		ID: "file.quickOpen", Title: "Go to File",
-		Handler: func() { openPalette(true) },
-	})
-
-	reg.Register(command.Command{
-		ID: "editor.goToLine", Title: "Go to Line",
-		Handler: func() { openPalette(false, ":") },
-	})
-
-	reg.Register(command.Command{
-		ID: "theme.switch", Title: "Switch Theme",
-		Handler: func() {
-			names := config.ListThemes()
-			if len(names) == 0 {
-				return
-			}
-			var cmds []command.Command
-			for _, name := range names {
-				cmds = append(cmds, command.Command{ID: name, Title: name})
-			}
-			picker := ui.NewCommandPaletteWidget(cmds)
-			picker.Borders = app.borders
-			originalStyleMap := app.screen.GetStyleMap()
-			originalPalette := *app.palette
-			applyTheme := func(theme config.ThemeConfig) {
-				app.screen.SetStyleMap(buildStyleMap(theme))
-				*app.palette = buildTerminalPalette(theme)
-				app.renderer.Clear()
-			}
-			picker.OnSelectionChange = func(name string) {
-				theme, err := config.LoadTheme(name)
-				if err != nil {
-					return
-				}
-				applyTheme(theme)
-			}
-			picker.OnExecute = func(name string) {
-				app.DismissDialog()
-				theme, err := config.LoadTheme(name)
-				if err != nil {
-					return
-				}
-				applyTheme(theme)
-				app.settings.Theme = name
-				config.SaveSettings(*app.settings)
-			}
-			picker.OnDismiss = func() {
-				app.DismissDialog()
-				app.screen.SetStyleMap(originalStyleMap)
-				*app.palette = originalPalette
-				app.renderer.Clear()
-			}
-			app.ShowDialog(picker)
-		},
-	})
-
-	openIndentPicker := func() {
-		var cmds []command.Command
-		sizes := []int{1, 2, 3, 4, 6, 8}
-		for _, s := range sizes {
-			label := fmt.Sprintf("Spaces: %d", s)
-			cmds = append(cmds, command.Command{ID: strconv.Itoa(s), Title: label})
-		}
-		cmds = append(cmds, command.Command{ID: "tabs", Title: "Indent Using Tabs"})
-		cmds = append(cmds, command.Command{ID: "detect", Title: "Detect from Content"})
-		app.ShowPicker(cmds, func(id string) {
-			if id == "detect" {
-				if app.editorGroup.Editor != nil && app.editorGroup.Editor.Buf != nil {
-					if info := buffer.DetectIndent(app.editorGroup.Editor.Buf.Lines); info.Size > 0 {
-						app.editorGroup.SetTabSize(info.Size)
-					}
-				}
-			} else if id == "tabs" {
-				app.editorGroup.SetTabSize(4)
-			} else if size, err := strconv.Atoi(id); err == nil {
-				app.editorGroup.SetTabSize(size)
-			}
-		})
-	}
-
-	app.statusBar.OnIndentClick = openIndentPicker
-
-	reg.Register(command.Command{
-		ID: "editor.indentation", Title: "Change Indentation",
-		Handler: openIndentPicker,
-	})
-
-	reg.Register(command.Command{
-		ID: "settings.open", Title: "Preferences: Open Settings",
-		Handler: func() {
-			path := config.ConfigFilePath("settings.json")
-			app.editorGroup.OpenFile(path)
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "keybindings.open", Title: "Preferences: Open Keyboard Shortcuts",
-		Handler: func() {
-			path := config.ConfigFilePath("keybindings.json")
-			app.editorGroup.OpenFile(path)
-		},
-	})
-}
-
-func registerExplorerCommands(reg *command.Registry, app *App) {
-	reg.Register(command.Command{
-		ID: "explorer.refresh", Title: "Refresh Explorer",
-		Handler: func() { app.explorer.Reload() },
-	})
-
-	reg.Register(command.Command{
-		ID: "explorer.open", Title: "Open",
-		Handler: func() {
-			app.explorer.ActivateSelected()
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "explorer.newFile", Title: "New File",
-		Handler: func() {
-			node := app.explorer.SelectedNode()
-			if node == nil {
-				return
-			}
-			parentDir := node.Path
-			if !node.IsDir {
-				parentDir = filepath.Dir(node.Path)
-			}
-			app.ShowInputDialog("New File", "Filename", "", func(name string) {
-				newPath := filepath.Join(parentDir, name)
-				if err := os.MkdirAll(filepath.Dir(newPath), 0755); err != nil {
-					app.StatusError("Error: " + err.Error())
-					return
-				}
-				if err := os.WriteFile(newPath, []byte{}, 0644); err != nil {
-					app.StatusError("Error: " + err.Error())
-					return
-				}
-				app.explorer.Reload()
-				app.editorGroup.OpenFile(newPath)
-				app.FocusEditor()
-			})
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "explorer.newFolder", Title: "New Folder",
-		Handler: func() {
-			node := app.explorer.SelectedNode()
-			if node == nil {
-				return
-			}
-			parentDir := node.Path
-			if !node.IsDir {
-				parentDir = filepath.Dir(node.Path)
-			}
-			app.ShowInputDialog("New Folder", "Folder name", "", func(name string) {
-				newPath := filepath.Join(parentDir, name)
-				if err := os.MkdirAll(newPath, 0755); err != nil {
-					app.StatusError("Error: " + err.Error())
-					return
-				}
-				app.explorer.Reload()
-			})
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "explorer.rename", Title: "Rename",
-		Handler: func() {
-			node := app.explorer.SelectedNode()
-			if node == nil {
-				return
-			}
-			app.ShowInputDialog("Rename", "New name", node.Name, func(newName string) {
-				dir := filepath.Dir(node.Path)
-				newPath := filepath.Join(dir, newName)
-				if err := os.Rename(node.Path, newPath); err != nil {
-					app.StatusError("Error: " + err.Error())
-					return
-				}
-				app.explorer.Reload()
-			})
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "explorer.delete", Title: "Delete",
-		Handler: func() {
-			node := app.explorer.SelectedNode()
-			if node == nil {
-				return
-			}
-			app.ShowConfirmDialog("Delete "+node.Name+"?",
-				[]string{"Yes", "No"},
-				[]func(){
-					func() {
-						app.DismissDialog()
-						if err := os.RemoveAll(node.Path); err != nil {
-							app.StatusError("Error: " + err.Error())
-							return
-						}
-						app.explorer.Reload()
-					},
-					func() { app.DismissDialog() },
-				},
-			)
-		},
-	})
-}
-
-func registerGitCommands(reg *command.Registry, app *App) {
-	reg.Register(command.Command{
-		ID: "changes.openDiff", Title: "Open Diff",
-		Handler: func() {
-			dir, status, ok := app.changes.SelectedFile()
-			if ok && app.changes.OnOpenDiff != nil {
-				app.changes.OnOpenDiff(dir, status)
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "changes.openFile", Title: "Open File",
-		Handler: func() {
-			fullPath := app.changes.SelectedFullPath()
-			if fullPath != "" {
-				app.editorGroup.OpenFile(fullPath)
-				app.root.SetFocus(app.editorGroup)
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "changes.refresh", Title: "Refresh Changes",
-		Handler: func() { app.changes.Refresh() },
-	})
-
-	reg.Register(command.Command{
-		ID: "changes.stage", Title: "Stage File",
-		Handler: func() {
-			dir, status, ok := app.changes.SelectedFile()
-			if ok && !status.Staged {
-				git.Stage(dir, status.Path)
-				app.changes.Refresh()
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "changes.unstage", Title: "Unstage File",
-		Handler: func() {
-			dir, status, ok := app.changes.SelectedFile()
-			if ok && status.Staged {
-				git.Unstage(dir, status.Path)
-				app.changes.Refresh()
-			}
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "changes.discard", Title: "Discard Changes",
-		Handler: func() {
-			dir, status, ok := app.changes.SelectedFile()
-			if ok && !status.Staged {
-				msg := fmt.Sprintf("Discard changes to %s? This is irreversible.", status.Path)
-				if status.Status == "?" {
-					msg = fmt.Sprintf("Delete untracked file %s? This is irreversible.", status.Path)
-				}
-				app.ShowConfirmDialog(msg,
-					[]string{"Cancel", "Discard"},
-					[]func(){
-						func() {
-							app.DismissDialog()
-						},
-						func() {
-							app.DismissDialog()
-							if status.Status == "?" {
-								git.DiscardUntracked(dir, status.Path)
-							} else {
-								git.Discard(dir, status.Path)
-							}
-							app.changes.Refresh()
-						},
-					},
-				)
-			}
-		},
-	})
-
-	registerGitCmd := func(id, title string, ops []func(string) error, verb string) {
-		reg.Register(command.Command{
-			ID: id, Title: title,
-			Handler: func() {
-				for _, dir := range app.changes.Dirs {
-					for _, op := range ops {
-						if err := op(dir); err != nil {
-							app.StatusError(fmt.Sprintf("%s failed: %v", verb, err))
-							return
-						}
-					}
-				}
-				app.StatusNotify(verb + " successfully")
-				app.changes.Refresh()
-			},
-		})
-	}
-	registerGitCmd("git.pull", "Git Pull", []func(string) error{git.Pull}, "Pulled")
-	registerGitCmd("git.push", "Git Push", []func(string) error{git.Push}, "Pushed")
-	registerGitCmd("git.sync", "Git Sync", []func(string) error{git.Pull, git.Push}, "Synced")
-}
-
-func registerWorkspaceCommands(reg *command.Registry, app *App) {
-	reg.Register(command.Command{
-		ID: "workspace.addFolder", Title: "Add Folder to Workspace",
-		Handler: func() {
-			app.ShowInputDialog("Add Folder", "Folder path", "", func(path string) {
-				if path == "" {
-					return
-				}
-				abs, err := filepath.Abs(path)
-				if err != nil {
-					app.StatusError("Error: " + err.Error())
-					return
-				}
-				info, err := os.Stat(abs)
-				if err != nil || !info.IsDir() {
-					app.StatusError("Not a directory: " + abs)
-					return
-				}
-				app.workspace.AddFolder(abs)
-				app.refreshWorkspaceWidgets()
-			})
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "workspace.removeFolder", Title: "Remove Folder from Workspace",
-		Handler: func() {
-			paths := app.workspace.Paths()
-			if len(paths) <= 1 {
-				app.StatusWarn("Cannot remove the last folder")
-				return
-			}
-			var cmds []command.Command
-			for _, p := range paths {
-				cmds = append(cmds, command.Command{ID: p, Title: filepath.Base(p)})
-			}
-			app.ShowPicker(cmds, func(path string) {
-				app.workspace.RemoveFolder(path)
-				app.refreshWorkspaceWidgets()
-			})
-		},
-	})
-
-	reg.Register(command.Command{
-		ID: "workspace.saveAs", Title: "Save Workspace As...",
-		Handler: func() {
-			app.ShowInputDialog("Save Workspace", "Filename", "workspace.ttt", func(path string) {
-				if path == "" {
-					return
-				}
-				if err := app.workspace.SaveFile(path); err != nil {
-					app.StatusError("Error: " + err.Error())
-				} else {
-					app.StatusNotify("Workspace saved: " + path)
-				}
-			})
-		},
-	})
-}
-
-func registerPRCommands(reg *command.Registry, app *App) {
-	reg.Register(command.Command{
-		ID: "pr.open", Title: "Open Pull Request",
-		Handler: func() {
-			if !github.IsGHInstalled() {
-				app.StatusError("GitHub CLI (gh) is required. Install from https://cli.github.com/")
-				return
-			}
-			dialog := ui.NewInputDialogWidget("Open Pull Request", "https://github.com/owner/repo/pull/123", "")
-			dialog.ConfirmLabel = "Open"
-			dialog.Borders = app.borders
-			dialog.OnSubmit = func(url string) {
-				app.DismissDialog()
-				if url != "" {
-					app.fetchAndOpenPR(url)
-				}
-			}
-			dialog.OnDismiss = func() {
-				app.DismissDialog()
-			}
-			app.ShowDialog(dialog)
-		},
-	})
-	reg.Register(command.Command{
-		ID: "pr.close", Title: "Close Pull Request",
-		Handler: func() {
-			app.changes.RemovePRGroups()
-		},
-	})
-}
-
-func registerWidgetCallbacks(reg *command.Registry, app *App) {
-	for i := range menuBarMenus {
-		idx := i
-		reg.Register(command.Command{
-			ID:    menuBarLabels[idx],
-			Title: "Open " + menuBarLabels[idx] + " Menu",
-			Handler: func() {
-				openMenuBarDropdown(app, reg, idx)
-			},
-		})
-	}
-
-	app.menuBar.OnSelect = func(index int) {
-		openMenuBarDropdown(app, reg, index)
-	}
-
-	app.root.OnRightClick = func(mx, my int) {
-		handleRightClick(app, reg, mx, my)
-	}
-
-	app.splitPanel.OnLeftClick = func() {
-		reg.Execute("sidebar.focus")
-	}
-	app.splitPanel.OnRightClick = func() {}
-
-
-	app.sidebar.MoreButton.OnClick = func(sx, sy int) {
-		var items []ui.ContextMenuItem
-		switch app.sidebar.ActivePanel {
-		case "explorer":
-			items = []ui.ContextMenuItem{
-				{Label: "New File", Command: "file.new"},
-				{Label: "Add Folder", Command: "workspace.addFolder"},
-				{Label: "Refresh", Command: "explorer.refresh"},
-			}
-		case "search":
-			replaceLabel := "Replace"
-			if app.search.IsReplaceMode() {
-				replaceLabel = "Search"
-			}
-			items = []ui.ContextMenuItem{
-				{Label: replaceLabel, Shortcut: app.KeyFor("sidebar.searchReplace"), Command: "sidebar.searchReplace"},
-				ui.MenuSep(),
-				{Label: "Clear Results", Command: "search.clear"},
-			}
-		case "changes":
-			items = []ui.ContextMenuItem{
-				{Label: "Refresh", Command: "changes.refresh"},
-				{Label: "Open Pull Request", Command: "pr.open"},
-				ui.MenuSep(),
-				{Label: "Pull", Command: "git.pull"},
-				{Label: "Push", Command: "git.push"},
-				{Label: "Sync", Command: "git.sync"},
-			}
-		}
-		if len(items) > 0 {
-			openContextMenu(app, reg, items, sx, sy)
-		}
-	}
-
-	app.sidebar.OnPanelChange = func(id string) {
-		if id == "search" {
-			app.applySearchHighlights()
-		} else {
-			app.editorGroup.ClearSearch()
-		}
-		if id == "changes" {
-			app.changes.Refresh()
-		}
-	}
-
-	app.sidebar.OnTabOverflow = func(ids []string, titles []string, sx, sy int) {
-		var items []ui.ContextMenuItem
-		for i, id := range ids {
-			panelID := id
-			items = append(items, ui.ContextMenuItem{Label: titles[i], Command: "sidebar.overflow." + panelID})
-			reg.Register(command.Command{
-				ID:      "sidebar.overflow." + panelID,
-				Title:   titles[i],
-				Handler: func() { app.sidebar.SetActivePanel(panelID) },
-			})
-		}
-		openContextMenu(app, reg, items, sx, sy)
-	}
-
-	app.editorGroup.TabBar.OnTabClose = func(index int) {
-		app.editorGroup.SwitchTab(index)
-		reg.Execute("tab.close")
-	}
-
-	app.editorGroup.TabBar.MoreButton.OnClick = func(sx, sy int) {
-		moreMenu := []ui.ContextMenuItem{
-			{Label: "Close All", Command: "tab.closeAll"},
-		}
-		openContextMenu(app, reg, moreMenu, sx, sy)
-	}
-
-	app.editorGroup.TabBar.OnTabRightClick = func(index, sx, sy int) {
-		app.editorGroup.SwitchTab(index)
-		tabContextMenu := []ui.ContextMenuItem{
-			{Label: "Close", Shortcut: app.KeyFor("tab.close"), Command: "tab.close"},
-			{Label: "Close Others", Shortcut: "", Command: "tab.closeOthers"},
-			{Label: "Close All", Shortcut: "", Command: "tab.closeAll"},
-		}
-		openContextMenu(app, reg, tabContextMenu, sx, sy)
-	}
-
-	app.explorer.OnOpenFile = func(path string) {
-		app.editorGroup.OpenFile(path)
-		app.root.SetFocus(app.editorGroup)
-	}
-
-	app.search.OnClear = func() {
-		app.editorGroup.ClearSearch()
-	}
-	app.search.PostEvent = func() {
-		app.screen.PostEvent(tcell.NewEventInterrupt(nil))
-	}
-	app.search.DiffSources = func() []ui.DiffSearchSource {
-		seen := map[string]bool{}
-		sources := app.editorGroup.DiffTabSources()
-		for _, s := range sources {
-			seen[s.TabName] = true
-		}
-		for _, g := range app.changes.Groups {
-			if !g.IsPR {
-				continue
-			}
-			for path, diffText := range g.PRDiffs {
-				tabName := path + " (diff)"
-				if seen[tabName] {
-					continue
-				}
-				fd := diff.Parse(diffText)
-				dv := ui.NewDiffViewWidget(path, fd)
-				sources = append(sources, ui.DiffSearchSource{TabName: tabName, Lines: dv.CombinedLines()})
-			}
-		}
-		return sources
-	}
-
-	app.search.OnOpenMatch = func(path string, line, col int) {
-		if strings.HasSuffix(path, " (diff)") {
-			if !app.editorGroup.SwitchToTabByPath(path) {
-				filePath := strings.TrimSuffix(path, " (diff)")
-				for _, g := range app.changes.Groups {
-					if !g.IsPR {
-						continue
-					}
-					if diffText, ok := g.PRDiffs[filePath]; ok {
-						app.editorGroup.OpenDiff(filePath, diff.Parse(diffText))
-						break
-					}
-				}
-			}
-			if dv := app.editorGroup.ActiveDiffWidget(); dv != nil {
-				dv.ScrollToLine(line - 1)
-				dv.ApplySearchHighlight(app.search.Input.Text, app.search.Options)
-			}
-			app.root.SetFocus(app.editorGroup)
-			return
-		}
-		app.editorGroup.OpenFile(path)
-		app.editorGroup.GoToLine(line)
-		if app.search.Input.Text != "" {
-			matches, _ := ui.FindInLines(app.editorGroup.Editor.Buf.Lines, app.search.Input.Text, app.search.Options)
-			app.editorGroup.SetSearch(app.search.Input.Text, matches)
-		}
-		app.root.SetFocus(app.editorGroup)
-	}
-
-	app.search.OnPreview = func(filePath string, matches []ui.SearchMatch, replacement string, opts ui.SearchOptions) {
-		data, err := os.ReadFile(filePath)
-		if err != nil {
-			app.StatusWarn("Cannot read file: " + err.Error())
-			return
-		}
-		lines := strings.Split(string(data), "\n")
-		if len(lines) > 0 && lines[len(lines)-1] == "" {
-			lines = lines[:len(lines)-1]
-		}
-		fd := ui.BuildReplaceDiff(filepath.Base(filePath), lines, matches, replacement, opts)
-		app.editorGroup.OpenDiff(filePath, fd)
-		app.root.SetFocus(app.editorGroup)
-	}
-
-	app.search.OnReplace = func(filePath string, matches []ui.SearchMatch, replacement string, opts ui.SearchOptions) {
-		data, err := os.ReadFile(filePath)
-		if err != nil {
-			app.StatusWarn("Cannot read file: " + err.Error())
-			return
-		}
-		lines := strings.Split(string(data), "\n")
-		if len(lines) > 0 && lines[len(lines)-1] == "" {
-			lines = lines[:len(lines)-1]
-		}
-		newLines := ui.ApplyReplacements(lines, matches, replacement, opts)
-		if err := os.WriteFile(filePath, []byte(strings.Join(newLines, "\n")+"\n"), 0644); err != nil {
-			app.StatusWarn("Cannot write file: " + err.Error())
-			return
-		}
-		app.editorGroup.ReloadFile(filePath)
-		app.search.Refresh()
-		app.StatusNotify(fmt.Sprintf("Replaced %d matches in %s", len(matches), filepath.Base(filePath)))
-	}
-
-	app.search.OnReplaceAll = func(allMatches map[string][]ui.SearchMatch, replacement string, opts ui.SearchOptions) {
-		totalFiles := len(allMatches)
-		totalMatches := 0
-		for _, m := range allMatches {
-			totalMatches += len(m)
-		}
-		msg := fmt.Sprintf("Replace %d matches across %d files? This cannot be undone.", totalMatches, totalFiles)
-		app.ShowConfirmDialog(msg, []string{"Cancel", "Replace All"}, []func(){
-			func() { app.DismissDialog() },
-			func() {
-				app.DismissDialog()
-				for filePath, matches := range allMatches {
-					data, err := os.ReadFile(filePath)
-					if err != nil {
-						continue
-					}
-					lines := strings.Split(string(data), "\n")
-					if len(lines) > 0 && lines[len(lines)-1] == "" {
-						lines = lines[:len(lines)-1]
-					}
-					newLines := ui.ApplyReplacements(lines, matches, replacement, opts)
-					if err := os.WriteFile(filePath, []byte(strings.Join(newLines, "\n")+"\n"), 0644); err != nil {
-						continue
-					}
-					app.editorGroup.ReloadFile(filePath)
-				}
-				app.search.Refresh()
-				app.StatusNotify(fmt.Sprintf("Replaced %d matches across %d files", totalMatches, totalFiles))
-			},
-		})
-	}
-
-	app.explorer.OnRightClick = func(node *ui.TreeNode, sx, sy int) {
-		items := []ui.ContextMenuItem{
-			{Label: "Open", Command: "explorer.open"},
-			ui.MenuSep(),
-			{Label: "New File", Command: "explorer.newFile"},
-			{Label: "New Folder", Command: "explorer.newFolder"},
-			ui.MenuSep(),
-			{Label: "Rename", Command: "explorer.rename"},
-			{Label: "Delete", Command: "explorer.delete"},
-		}
-		openContextMenu(app, reg, items, sx, sy)
-	}
-
-	app.changes.OnRightClick = func(dir string, status git.FileStatus, sx, sy int) {
-		if status.Staged {
-			openContextMenu(app, reg, changesContextMenuStaged, sx, sy)
-		} else {
-			openContextMenu(app, reg, changesContextMenuUnstaged, sx, sy)
-		}
-	}
-
-	app.changes.OnOpenDiff = func(dir string, status git.FileStatus) {
-		fullPath := filepath.Join(dir, status.Path)
-		if status.Status == "?" {
-			app.editorGroup.OpenFile(fullPath)
-			app.root.SetFocus(app.editorGroup)
-			return
-		}
-		var diffText string
-		var err error
-		if status.Status == "R" && status.OldPath != "" {
-			diffText, err = git.DiffRename(dir, status.OldPath, status.Path)
-		} else {
-			diffText, err = git.DiffFile(dir, status.Path)
-		}
-		if err != nil || diffText == "" {
-			app.editorGroup.OpenFile(fullPath)
-			app.root.SetFocus(app.editorGroup)
-			return
-		}
-		parsed := diff.Parse(diffText)
-		if len(parsed.Hunks) == 0 {
-			app.editorGroup.OpenFile(fullPath)
-			app.root.SetFocus(app.editorGroup)
-			return
-		}
-		app.editorGroup.OpenDiff(status.Path, parsed)
-		app.root.SetFocus(app.editorGroup)
-	}
-
-	app.changes.OnOpenPRDiff = func(group *ui.ChangesGroup, status git.FileStatus) {
-		diffText, ok := group.PRDiffs[status.Path]
-		if !ok || diffText == "" {
-			app.StatusWarn("No diff available for " + status.Path)
-			return
-		}
-		parsed := diff.Parse(diffText)
-		if len(parsed.Hunks) == 0 {
-			app.StatusWarn("Empty diff for " + status.Path)
-			return
-		}
-		app.editorGroup.OpenDiff(status.Path, parsed)
-		app.root.SetFocus(app.editorGroup)
-	}
-
-	app.changes.OnPRGroupMenu = func(group *ui.ChangesGroup, sx, sy int) {
-		name := group.Name
-		url := group.PRURL
-		refreshID := "pr.refresh." + name
-		closeID := "pr.close." + name
-		reg.Register(command.Command{
-			ID: refreshID, Title: "Refresh",
-			Handler: func() {
-				app.changes.RemovePRGroup(name)
-				app.fetchAndOpenPR(url)
-			},
-		})
-		reg.Register(command.Command{
-			ID: closeID, Title: "Close",
-			Handler: func() {
-				app.changes.RemovePRGroup(name)
-			},
-		})
-		items := []ui.ContextMenuItem{
-			{Label: "Refresh", Command: refreshID},
-			{Label: "Close", Command: closeID},
-		}
-		openContextMenu(app, reg, items, sx, sy)
-	}
-
-	app.changes.OnGroupMenu = func(dir string, sx, sy int) {
-		items := []ui.ContextMenuItem{
-			{Label: "Pull", Command: "git.pull." + dir},
-			{Label: "Push", Command: "git.push." + dir},
-			{Label: "Sync", Command: "git.sync." + dir},
-		}
-		registerDirGitCmd := func(id, title string, ops []func(string) error, verb string) {
-			reg.Register(command.Command{
-				ID: id, Title: title,
-				Handler: func() {
-					for _, op := range ops {
-						if err := op(dir); err != nil {
-							app.StatusError(fmt.Sprintf("%s failed: %v", verb, err))
-							return
-						}
-					}
-					app.StatusNotify(verb + " successfully")
-					app.changes.Refresh()
-				},
-			})
-		}
-		registerDirGitCmd("git.pull."+dir, "Pull", []func(string) error{git.Pull}, "Pulled")
-		registerDirGitCmd("git.push."+dir, "Push", []func(string) error{git.Push}, "Pushed")
-		registerDirGitCmd("git.sync."+dir, "Sync", []func(string) error{git.Pull, git.Push}, "Synced")
-		openContextMenu(app, reg, items, sx, sy)
-	}
-
-	app.changes.OnCommit = func(dir string, message string) {
-		if err := git.Commit(dir, message); err != nil {
-			app.StatusError("Commit failed: " + err.Error())
-		} else {
-			for i := range app.changes.Groups {
-				if app.changes.Groups[i].Dir == dir {
-					app.changes.Groups[i].Input.Clear()
-				}
-			}
-			app.StatusNotify("Committed: " + message)
-			app.changes.Refresh()
-		}
-	}
-
-	app.changes.OnConfirmDiscard = func(message string, onConfirm func()) {
-		app.ShowConfirmDialog(message,
-			[]string{"Cancel", "Discard"},
-			[]func(){
-				func() {
-					app.DismissDialog()
-				},
-				func() {
-					app.DismissDialog()
-					onConfirm()
-				},
-			},
-		)
-	}
-
-	app.contentSplit.OnResize = func(height int) {
-		if height <= 0 {
-			app.contentSplit.ShowBottom = false
-		} else {
-			app.contentSplit.ShowBottom = true
-			app.contentSplit.BottomH = height
-			if len(app.terminals) == 0 {
-				app.SpawnTerminal()
-			}
-		}
-	}
-
-	app.contentSplit.OnTopClick = func() {
-		app.root.SetFocus(app.editorGroup)
-	}
-
-	app.contentSplit.OnBottomClick = func() {
-		if w := app.bottomPanel.ActiveWidget(); w != nil {
-			app.root.SetFocus(w)
-		}
-	}
-
-	app.splitPanel.OnResize = func(width int) {
-		app.SetSidebarWidth(width)
-	}
-
-	app.bottomPanel.TabBar.OnTabClick = func(index int) {
-		panels := app.bottomPanel.PanelIDs()
-		if index >= 0 && index < len(panels) {
-			app.bottomPanel.SetActivePanel(panels[index])
-			if w := app.bottomPanel.ActiveWidget(); w != nil {
-				app.root.SetFocus(w)
-			}
-		}
-	}
-
-	app.bottomPanel.TabBar.OnAdd = func() {
-		reg.Execute("terminal.new")
-	}
-
-	app.bottomPanel.TabBar.MoreButton = ui.NewMoreButtonWidget()
-	app.bottomPanel.TabBar.MoreButton.OnClick = func(sx, sy int) {
-		items := []ui.ContextMenuItem{
-			{Label: "New Terminal", Command: "terminal.new"},
-			ui.MenuSep(),
-			{Label: "Close All Terminals", Command: "terminal.closeAll"},
-		}
-		openContextMenu(app, reg, items, sx, sy)
-	}
-}
-
 
 func bindKeys(root *ui.Root, reg *command.Registry, keybindings []config.KeyBinding) {
 	for _, kb := range keybindings {
